+++ assets/app.stories-D7pzEp3a.js
@@ -1,0 +1,80 @@
+import{j as e}from"./iframe-BHBhYjQq.js";import{M as p}from"./index-BepVl9yJ.js";import{S as c}from"./index-BKvPWbv6.js";import{A as t,M as d,C as l}from"./MainLayout-CWFMT-HR.js";import{O as v}from"./OverviewPage-CLX1x8DM.js";import{H as u}from"./HomeIcon-GaB_KH5Q.js";import"./FilledIcon-Cz3BrzVq.js";import"./DropdownIcon-BFI0RSoQ.js";import"./type-Bc9hys3F.js";import"./Heading-DoIhT-Cb.js";import"./AddIcon-CR4vMKGb.js";import"./useOutsideToggle-BXaoZqFj.js";import"./DiviCustomFieldHeplerIcon-BbRs-AaR.js";import"./DiviDynamicHelperIcon-x9BIyvjO.js";import"./PACLogo-CPFDbwrn.js";import"./index-dSBykjbi.js";import"./index-DT2JD236.js";import"./index-D02EE_ej.js";import"./CloseDropDownIcon-DjUVijBa.js";import"./ChevronRightIcon-Cgb-mamO.js";import"./BackIcon-EpUDms-H.js";import"./index-hfYVbx_S.js";import"./index-Q1_gUDhq.js";import"./Text-_q3N7_9X.js";import"./index-CSJmvfyY.js";import"./index-HzealvQc.js";import"./CloseIcon-B7sQNzeI.js";import"./KeyIcon-DeHNTin8.js";import"./DocumentationIcon-BLPD9wiK.js";import"./OpenIcon-C6zC8tmz.js";import"./SupportIcon-k8h76ZwE.js";import"./UsersIcon-DwrswyIR.js";import"./index-D4B-LMGz.js";import"./EyeIcon-4JaCchU4.js";import"./index-D-hpkK-X.js";import"./index-C9SILz9g.js";import"./index-B5518ZIg.js";import"./index-Dq7h7Pqt.js";import"./index-CIksEH5x.js";import"./InfoIcon-D9ciRe7f.js";import"./CheckmarkIcon-BO1SBL60.js";import"./StopIcon-DvKS5CvV.js";import"./TextLink-DI0ReQ0Q.js";const oe={title:"PAC Universal Plugin UI/Layout/App",componnet:t,parameters:{pageLayout:"page"}},i=()=>{const r={avatar:"https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",name:"Muhammad Touseef",greeting:"Good afternoon",pluginLogo:"https://dummyjson.com/image/74",pluginDescription:"A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",pluginName:"Divi Assistant",pacPluginNewsApi:"https://dummyjson.com/c/f371-88b4-4d1e-b88e",onPacCardClick:()=>{console.log("PAC Card Clicked")},docLink:"https://www.peeayecreative.com/docs/divi-assistant/",supportLink:"https://www.peeayecreative.com/support/",communityLink:"https://www.peeayecreative.com/community/space/divi-assistant/home"},s={version:"0.0.9",userAvatar:"https://bc3-production-assets-cdn.basecamp-static.com/4677743/people/BAhpBJQn1wI=--4b34d6ad2aa165f177c7e08ebb0fbb25e7fb5da0/avatar/in/260/90ad41c3cfd4e50976d4a0fad1ef44c8.avif?v=1",userName:"Muhammad Touseef",activeMenu:"",plugins:[{id:"divi-assistant",name:"Divi Assistant"},{id:"divi-dynamic-helper",name:"Divi Dynamic Content Helper"},{id:"divi-contact-form-helper",name:"Divi Contact Form Helper"},{id:"divi-custom-fields-helper",name:"Divi Custom Fields Helper"},{id:"divi-events-calendar",name:"Divi Events Calendar"},{id:"divi-taxonomy-helper",name:"Divi Taxonomy Helper"},{id:"divi-search-helper",name:"Divi Search Helper"},{id:"divi-table-of-contents-maker",name:"Divi Table of Contents Maker"},{id:"divi-carousel-maker",name:"Divi Carousel Maker"},{id:"divi-responsive-helper",name:"Divi Responsive Helper"},{id:"divi-social-sharing-buttons-maker",name:"Divi Social Sharing Buttons Maker"},{id:"divi-timer-pro",name:"Divi Timer Pro"},{id:"divi-tabs-maker",name:"Divi Tabs Maker"}],activePluginId:"divi-responsive-helper",onPluginSelect:()=>{},onAvatarClick:()=>console.log("Go to Admin Page")},m=[{id:"overview",label:"Overview",icon:u}];return e.jsxs(t,{height:"400px",children:[e.jsx(p,{...s}),e.jsxs(d,{children:[e.jsx(c,{menuData:m,activeMenu:"overview",onActiveSubmenuChange:()=>{},onActiveMenuChange:()=>{},showAddNewMenu:!0,showAllPluginMenu:!0}),e.jsx(l,{children:e.jsx(v,{...r})})]})]})};i.__docgenInfo={description:"",methods:[],displayName:"App"};var n,a,o;i.parameters={...i.parameters,docs:{...(n=i.parameters)==null?void 0:n.docs,source:{originalSource:`() => {
+  const OverviewPageProps = {
+    avatar: "https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",
+    name: "Muhammad Touseef",
+    greeting: "Good afternoon",
+    pluginLogo: "https://dummyjson.com/image/74",
+    pluginDescription: "A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",
+    pluginName: "Divi Assistant",
+    pacPluginNewsApi: "https://dummyjson.com/c/f371-88b4-4d1e-b88e",
+    onPacCardClick: () => {
+      console.log("PAC Card Clicked");
+    },
+    docLink: "https://www.peeayecreative.com/docs/divi-assistant/",
+    supportLink: "https://www.peeayecreative.com/support/",
+    communityLink: "https://www.peeayecreative.com/community/space/divi-assistant/home"
+  };
+  const MainheaderProps = {
+    version: "0.0.9",
+    userAvatar: "https://bc3-production-assets-cdn.basecamp-static.com/4677743/people/BAhpBJQn1wI=--4b34d6ad2aa165f177c7e08ebb0fbb25e7fb5da0/avatar/in/260/90ad41c3cfd4e50976d4a0fad1ef44c8.avif?v=1",
+    userName: "Muhammad Touseef",
+    activeMenu: "",
+    plugins: [{
+      id: "divi-assistant",
+      name: "Divi Assistant"
+    }, {
+      id: "divi-dynamic-helper",
+      name: "Divi Dynamic Content Helper"
+    }, {
+      id: "divi-contact-form-helper",
+      name: "Divi Contact Form Helper"
+    }, {
+      id: "divi-custom-fields-helper",
+      name: "Divi Custom Fields Helper"
+    }, {
+      id: "divi-events-calendar",
+      name: "Divi Events Calendar"
+    }, {
+      id: "divi-taxonomy-helper",
+      name: "Divi Taxonomy Helper"
+    }, {
+      id: "divi-search-helper",
+      name: "Divi Search Helper"
+    }, {
+      id: "divi-table-of-contents-maker",
+      name: "Divi Table of Contents Maker"
+    }, {
+      id: "divi-carousel-maker",
+      name: "Divi Carousel Maker"
+    }, {
+      id: "divi-responsive-helper",
+      name: "Divi Responsive Helper"
+    }, {
+      id: "divi-social-sharing-buttons-maker",
+      name: "Divi Social Sharing Buttons Maker"
+    }, {
+      id: "divi-timer-pro",
+      name: "Divi Timer Pro"
+    }, {
+      id: "divi-tabs-maker",
+      name: "Divi Tabs Maker"
+    }],
+    activePluginId: "divi-responsive-helper",
+    onPluginSelect: () => {},
+    onAvatarClick: () => console.log("Go to Admin Page")
+  };
+  const sidemenuData = [{
+    id: "overview",
+    label: "Overview",
+    icon: HomeIcon
+  }];
+  return <AppLayout height="400px">
+      <MainHeader {...MainheaderProps} />
+      <MainLayout>
+        <Sidebar menuData={sidemenuData} activeMenu={"overview"} onActiveSubmenuChange={() => {}} onActiveMenuChange={() => {}} showAddNewMenu showAllPluginMenu />
+        <ContentLayout>
+          <OverviewPage {...OverviewPageProps} />
+        </ContentLayout>
+      </MainLayout>
+    </AppLayout>;
+}`,...(o=(a=i.parameters)==null?void 0:a.docs)==null?void 0:o.source}}};const te=["App"];export{i as App,te as __namedExportsOrder,oe as default};