+++ assets/icon.stories-CROzY9Pu.js
@@ -1,0 +1,46 @@
+import{j as n}from"./iframe-BHBhYjQq.js";import{A as I,C as p,G as r,S as i}from"./CacheIcon-D9SgKaNC.js";import{A as C,V as l,W as d}from"./WebsiteIcon-D3x-4Mt0.js";import{A as u}from"./AddIcon-CR4vMKGb.js";import{A as g,a as f,b as h,c as y,B as S,d as x,e as b,C as T,f as v,g as k,h as P,i as D,j as L,k as F,D as A,l as U,L as E,m as B,M,F as R,n as w,o as j,G as H,H as z,p as O,q as N,I as V,r as W,s as G,t as q,u as K,v as _,P as Q,w as X,x as Y,y as Z,N as J,z as $,E as nn,O as on,J as en,K as an,Q as mn,R as tn,S as cn,T as sn,U as In,V as pn,W as rn,X as Cn,Y as ln,Z as dn,_ as un,$ as gn,a0 as fn,a1 as hn,a2 as yn,a3 as Sn,a4 as xn,a5 as bn,a6 as Tn,a7 as vn,a8 as kn,a9 as Pn,aa as Dn,ab as Ln,ac as Fn,ad as An,ae as Un,af as En,ag as Bn,ah as Mn,ai as Rn,aj as wn,ak as jn,al as Hn,am as zn,an as On,ao as Nn,ap as Vn}from"./DragIcon-DgEfyRVg.js";import{B as Wn}from"./BackIcon-EpUDms-H.js";import{B as Gn,F as qn,K as Kn}from"./KeyIcon-DeHNTin8.js";import{C as _n,L as Qn}from"./LearningIcon-BrWtDAAL.js";import{C as Xn}from"./ChangelogIcon-XyWxXtCP.js";import{C as Yn}from"./CheckmarkIcon-BO1SBL60.js";import{C as Zn}from"./CloseIcon-B7sQNzeI.js";import{C as Jn,E as $n}from"./EditIcon-B8iztJkR.js";import{D as no}from"./DocumentationIcon-BLPD9wiK.js";import{E as oo}from"./EmailIcon-CE2THpnY.js";import{F as eo}from"./FaqIcon-BKqH2KaO.js";import{H as ao}from"./HomeIcon-GaB_KH5Q.js";import{I as mo}from"./ImportExportIcon-DNmAaVFG.js";import{I as to}from"./InfoIcon-D9ciRe7f.js";import{I as co,S as so,T as Io,U as po}from"./UnderlinedIcon-CWq57mJX.js";import{R as ro}from"./ResetIcon-WjjG4tPR.js";import{S as io,U as Co}from"./UpdatesIcon-BCjU3kRK.js";import{S as lo}from"./SearchIcon-CQ0mTVia.js";import{S as uo}from"./SettingsIcon-Csn-dAGb.js";import{S as go}from"./SupportIcon-k8h76ZwE.js";import{U as fo}from"./UsersIcon-DwrswyIR.js";import{E as ho}from"./EyeIcon-4JaCchU4.js";import{O as yo}from"./OpenIcon-C6zC8tmz.js";import{C as So}from"./ChevronRightIcon-Cgb-mamO.js";import{D as xo}from"./DropdownIcon-BFI0RSoQ.js";import{W as bo,S as To}from"./StopIcon-DvKS5CvV.js";import{C as vo}from"./CloseDropDownIcon-DjUVijBa.js";import"./type-Bc9hys3F.js";const ko=[{name:"AccessibilityIcon",Component:I},{name:"ActionIcon",Component:C},{name:"AddIcon",Component:u},{name:"AdminBarIcon",Component:g},{name:"AiIcon",Component:f},{name:"AltTextIcon",Component:h},{name:"AssignedIcon",Component:y},{name:"BackIcon",Component:Wn},{name:"BackToTopIcon",Component:S},{name:"BoldIcon",Component:x},{name:"BrowserIcon",Component:b},{name:"BulletListIcon",Component:Gn},{name:"CacheIcon",Component:p},{name:"CalendarIcon",Component:T},{name:"ChangeIcon",Component:_n},{name:"ChangelogIcon",Component:Xn},{name:"CheckmarkIcon",Component:Yn},{name:"ChildThemeIcon",Component:v},{name:"CloseIcon",Component:Zn},{name:"CloudUpIcon",Component:k},{name:"ColorsIcon",Component:P},{name:"ComingSoonIcon",Component:D},{name:"CopyIcon",Component:Jn},{name:"CustomCodeIcon",Component:L},{name:"CustomFieldsIcon",Component:F},{name:"TrashIcon",Component:A},{name:"DesktopIcon",Component:U},{name:"LibraryIcon",Component:E},{name:"DocumentationIcon",Component:no},{name:"DownloadIcon",Component:B},{name:"MenuIcon",Component:M},{name:"EditIcon",Component:$n},{name:"EmailIcon",Component:oo},{name:"FallbackIcon",Component:R},{name:"FaqIcon",Component:eo},{name:"FeatureRequestIcon",Component:qn},{name:"FontIcon",Component:w},{name:"FormsIcon",Component:j},{name:"GlossaryIcon",Component:r},{name:"GroupsIcon",Component:H},{name:"HeadingIcon",Component:z},{name:"HealthIcon",Component:O},{name:"HideIcon",Component:N},{name:"HomeIcon",Component:ao},{name:"ImportExportIcon",Component:mo},{name:"IncrementIcon",Component:V},{name:"InfoIcon",Component:to},{name:"InterfaceIcon",Component:W},{name:"ItalicIcon",Component:co},{name:"KeyIcon",Component:Kn},{name:"LearningIcon",Component:Qn},{name:"LinkIcon",Component:G},{name:"LoginIcon",Component:q},{name:"MaintenanceConeIcon",Component:K},{name:"MaintenanceIcon",Component:_},{name:"PauseIcon",Component:Q},{name:"ImageIcon",Component:X},{name:"MediaLibraryIcon",Component:Y},{name:"ModulesIcon",Component:Z},{name:"NotesIcon",Component:J},{name:"NotificationsIcon",Component:$},{name:"NumberListIcon",Component:nn},{name:"OwnersManualIcon",Component:on},{name:"ParagraphIcon",Component:en},{name:"PasswordIcon",Component:an},{name:"PhasesIcon",Component:mn},{name:"PhoneIcon",Component:tn},{name:"PluginListIcon",Component:cn},{name:"PinIcon",Component:sn},{name:"PresetsIcon",Component:In},{name:"PrintIcon",Component:pn},{name:"ProductsIcon",Component:rn},{name:"ReplyIcon",Component:Cn},{name:"ResetIcon",Component:ro},{name:"CheckCircleIcon",Component:ln},{name:"RolesIcon",Component:dn},{name:"SaveIcon",Component:io},{name:"ScrollbarIcon",Component:un},{name:"ScrollIcon",Component:gn},{name:"SearchIcon",Component:lo},{name:"SettingsIcon",Component:uo},{name:"ShareIcon",Component:fn},{name:"StartupIcon",Component:i},{name:"StatusIcon",Component:hn},{name:"StrikethroughIcon",Component:so},{name:"FilterIcon",Component:yn},{name:"SupportIcon",Component:go},{name:"TabletIcon",Component:Sn},{name:"TagIcon",Component:xn},{name:"TemplateIcon",Component:bn},{name:"TextIcon",Component:Io},{name:"ToggleIcon",Component:Tn},{name:"CloseCircleIcon",Component:vn},{name:"TroubleshootingIcon",Component:kn},{name:"UnderlinedIcon",Component:po},{name:"UpdatesIcon",Component:Co},{name:"UploadIcon",Component:Pn},{name:"DashboardCustomizeIcon",Component:Dn},{name:"UsersIcon",Component:fo},{name:"UtilityIcon",Component:Ln},{name:"VideoIcon",Component:l},{name:"EyeIcon",Component:ho},{name:"OpenIcon",Component:yo},{name:"ChevronRightIcon",Component:So},{name:"DropdownIcon",Component:xo},{name:"CloseDropDownIcon",Component:vo},{name:"WelcomeIcon",Component:bo},{name:"EditorIcon",Component:Fn},{name:"EnvironmentsIcon",Component:An},{name:"QuoteIcon",Component:Un},{name:"WebsiteIcon",Component:d},{name:"TooltipIcon",Component:En},{name:"SpamIcon",Component:Bn},{name:"ReplaceIcon",Component:Mn},{name:"PluginIcon",Component:Rn},{name:"MeasurementIcon",Component:wn},{name:"LocationIcon",Component:jn},{name:"ImageFavouriteIcon",Component:Hn},{name:"ImageCheckIcon",Component:zn},{name:"HighlightIcon",Component:On},{name:"DragIcon",Component:Nn},{name:"VariableIcon",Component:Vn},{name:"StopIcon",Component:To}],me={title:"PAC Universal Plugin UI/Icons/Outlined",parameters:{layout:"padded"},argTypes:{size:{description:"Size of the Link",control:{type:"select"},options:["xsmall","small","medium","large","xlarge"]},color:{description:"Color of the icon",control:{type:"color"},defaultValue:"var(--pac-upui-icon)"}},args:{size:"xlarge"}},o={render:e=>n.jsx("div",{style:{display:"grid",gridTemplateColumns:"repeat(5, 1fr)",gap:20,justifyContent:"center"},children:ko.map(({name:a,Component:s})=>n.jsxs("div",{style:{display:"flex",flexDirection:"column",alignItems:"center",justifyContent:"center",gap:"10px",border:"1px solid rgb(215, 218, 221)",padding:"6px",position:"relative"},children:[n.jsx("div",{style:{display:"flex"},children:n.jsx(s,{size:e.size,color:e.color})}),n.jsx("div",{children:a})]},a))})};var m,t,c;o.parameters={...o.parameters,docs:{...(m=o.parameters)==null?void 0:m.docs,source:{originalSource:`{
+  render: args => <div style={{
+    display: "grid",
+    gridTemplateColumns: "repeat(5, 1fr)",
+    gap: 20,
+    justifyContent: "center"
+  }}>
+      {icons.map(({
+      name,
+      Component
+    }) => <div key={name} style={{
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "center",
+      justifyContent: "center",
+      gap: "10px",
+      border: "1px solid rgb(215, 218, 221)",
+      padding: "6px",
+      position: "relative"
+    }}>
+          <div style={{
+        display: "flex"
+      }}>
+            <Component size={args.size} color={args.color} />
+          </div>
+          <div>{name}</div>
+          {/* <div
+            onClick={() => {
+                          navigator.clipboard.writeText("XYUZ").then(() => {
+                toast.success("Icon name copied to clipboard!");
+              }, () => {
+                toast.error("Failed to copy icon name.");
+              });
+            }}
+            style={{
+              position: "absolute",
+              top: "5px",
+              right: "5px",
+              cursor: "pointer",
+            }}
+           >
+            <CopyIcon />
+           </div> */}
+        </div>)}
+    </div>
+}`,...(c=(t=o.parameters)==null?void 0:t.docs)==null?void 0:c.source}}};const te=["Outlined"];export{o as Outlined,te as __namedExportsOrder,me as default};