+++ assets/index-CNUF2PB-.js
@@ -1,1 +1,0 @@
-import{g as r}from"./iframe-BdSb4ebm.js";import{r as o}from"./index-DV5lWZup.js";var t=o();const m=r(t);export{m as H,t as r};