+++ assets/introcard.stories-DbMePlFW.js
@@ -1,5 +1,0 @@
-import{j as t}from"./iframe-BdSb4ebm.js";import{I as s}from"./index-D0Diw0wC.js";import"./UsersIcon-B_fig-mV.js";import"./type-Bc9hys3F.js";import"./WebsiteIcon-u13GPJbk.js";import"./index-qpDThYTA.js";import"./index-CqyDz2mV.js";import"./index-BLdHY2cA.js";import"./OpenIcon-DBZbVxJG.js";import"./Heading-BR177-iI.js";import"./Text-iV-n9ZSW.js";const h={title:"PAC Universal Plugin UI/Composite/Intro Card",component:s},r={render:()=>t.jsx("div",{style:{width:"var(--pac-upui-width-sidebar)"},children:t.jsx(s,{})})};var o,i,e;r.parameters={...r.parameters,docs:{...(o=r.parameters)==null?void 0:o.docs,source:{originalSource:`{
-  render: () => <div style={{
-    width: "var(--pac-upui-width-sidebar)"
-  }}><IntroCard /></div>
-}`,...(e=(i=r.parameters)==null?void 0:i.docs)==null?void 0:e.source}}};const C=["Introduction"];export{r as Introduction,C as __namedExportsOrder,h as default};